feat(automation/regex): adds type extraction to regex
